+++ Extension/src/LanguageServer/client.ts
@@ -841,7 +841,7 @@ class DefaultClient implements Client {
         let decoration: vscode.TextEditorDecorationType = vscode.window.createTextEditorDecorationType({
             opacity: settings.inactiveRegionOpacity.toString(),
             backgroundColor: settings.inactiveRegionBackgroundColor,
-            color: settings.inactiveRegionFontColor,
+            color: settings.inactiveRegionForegroundColor,
             rangeBehavior: vscode.DecorationRangeBehavior.ClosedOpen
         });